fix_открытие соединения сокета, ProtectedRoute для попапа истории заказов
